Separate context custom hooks to avoid re-endering of components
