+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import { GlobalContextProvider } from './context/use-cocktailContext';
+import { GlobalContextProvider } from './context/use-context';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
